src/requirements.txt
aiohttp==3.8.0
aioredis==1.3.1
aiosignal==1.2.0
alabaster==0.7.12
amqp==2.6.1
argh==0.26.2
asgiref==3.4.1
astroid==2.8.4
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
bcrypt==3.2.0
billiard==3.6.4.0
boto3==1.17.109
botocore==1.20.109
celery==4.4.2
certifi==2021.10.8
cffi==1.14.6
channels==2.4.0
channels-redis==2.4.2
chardet==4.0.0
charset-normalizer==2.0.7
config==0.5.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==2.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0
django-3-jet==1.0.8
django-admin-rangefilter==0.3.15
django-amazon-ses==4.0.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==5.9.0
django-cleanup==5.1.0
django-compat==1.0.15
django-cors-headers==3.7.0
django-crum==0.7.5
django-defender==0.9.0
django-dynamic-fixture==3.1.1
django-environ==0.8.1
django-filter==2.2.0
django-geoposition-2==0.3.11
django-hijack==2.3.0
django-hijack-admin==2.1.10
django-import-export==2.6.1
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.11.0
django-rest-swagger==2.2.0
django-rosetta==0.9.4
django-solo==1.1.3
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.7.2
docutils==0.17.1
environ==1.0
et-xmlfile==1.1.0
Fabric3==1.14.post1
frozenlist==1.2.0
hiredis==1.0.1
hyperlink==21.0.0
idna==3.3
imagesize==1.2.0
importlib-metadata==4.8.1
incremental==21.3.0
isort==5.10.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.6.0
loducode-utils==1.5.1
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
multidict==5.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
paramiko==2.7.2
Pillow==7.0.0
platformdirs==2.4.0
polib==1.1.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.3
PyYAML==6.0
redis==3.4.1
requests==2.26.0
s3transfer==0.4.2
service-identity==21.1.0
simplejson==3.17.5
six==1.16.0
slackclient==2.9.3
snowballstemmer==2.1.0
Sphinx==4.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tablib==3.1.0
tblib==1.7.0
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
watchdog==2.1.3
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.6.0
zope.interface==5.4.0
